+++ src/config/default.docunotion.config.ts
@@ -3,13 +3,13 @@ import {
   imgurGifEmbed,
   vimeoEmbed,
   youtubeEmbed,
-} from "../plugins/StandardEmbeddings";
+} from "../plugins/embedTweaks";
 import { standardImageTransformer } from "../images";
 import { standardInternalLinkConversion } from "../plugins/internalLinks";
 import { standardCalloutTransformer } from "../plugins/CalloutTransformer";
 import { standardColumnListTransformer } from "../plugins/ColumnListTransformer";
 import { standardColumnTransformer } from "../plugins/ColumnTransformer";
-import { escapeHtmlBlockModifier } from "../plugins/EscapeHtmlBlockModifier";
+import { standardEscapeHtmlBlockModifier } from "../plugins/EscapeHtmlBlockModifier";
 import { standardHeadingTransformer } from "../plugins/HeadingTransformer";
 import { standardNumberedListTransformer } from "../plugins/NumberedListTransformer";
 import { standardTableTransformer } from "../plugins/TableTransformer";
@@ -19,7 +19,7 @@ import { standardExternalLinkConversion } from "../plugins/externalLinks";
 const defaultConfig: IDocuNotionConfig = {
   plugins: [
     // Notion "Block" JSON modifiers
-    escapeHtmlBlockModifier,
+    standardEscapeHtmlBlockModifier,
     standardHeadingTransformer, // does operations on both the Notion JSON and then later, on the notion to markdown transform
 
     // Notion to Markdown transformers. Most things get transformed correctly by the notion-to-markdown library,
