+++ src/plugins/embedTweaks.ts
